requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-queue==2.1.0
pyyaml==6.0
uritemplate==4.1.1
azure-storage-file-share
azure.storage.blob
fontawesomefree
pdf2image
pillow
azure-cognitiveservices-vision-computervision
azure-ai-formrecognizer